// src/utils/enums.ts
export enum Roles {
  Leads = 'LEADS',
  Customer = 'CUSTOMER',
  FieldAgent = 'FIELDAGENT',
  UnderWriter = 'UNDERWRITER',
  Manager = 'MANAGER',
  Admin = 'ADMIN',
  FinanceManager = 'FINANCEMANAGER'
}

export enum FundingFromCurrentStatus {
  Inprogress = 'Inprogress',
  Submitted = 'Submitted',
  AgentSubmitted = 'Agent_Submitted',
  UnderwriterSubmitted = 'Underwriter_Submitted',
  ManagerApproved = 'Manager_Approved',
  AdminCashDisbursed = 'Admin_Cash_Disbursed',
  UnderwriterReturned = 'Underwriter_Returned',
  ManagerRejected = 'Manager_Rejected',
  AdminRejected = 'Admin_Rejected',
  AmountCredited = 'Amount_Credited',
  Completed = 'Completed',
  MovedToLegal = 'Moved_To_Legal',
  FundingClosed = 'Funding_Closed'
}
// ...
export enum contractStatus {
  processing = 'processing',
  sent = 'sent',
  resent = 'resent',
  opened = 'opened',
  signed = 'signed',
  signedByAll = 'signed_by_all',
  cancelled = 'cancelled',
  expired = 'expired',
  failed = 'failed',
  draft = 'draft'
}